server/lat_long_API_server/requirements.txt
grpcio
protobuf
requests
python-dotenv